+++ src/uqm/supermelee/melee.cpp
@@ -503,7 +503,7 @@ void DrawShipPickerText(STAMP stamp)
 	pt.y = r.corner.y + SP_Y_PADDING;
 
 	s.frame = SetAbsFrameIndex(MeleeFrame,
-							   CONFIRM_PC + (optControllerType * 4));
+							   CONFIRM_PC + (static_cast<int>(optControllerType) * 4));
 	s.origin = pt;
 
 	DrawStamp(&s);
@@ -521,7 +521,7 @@ void DrawShipPickerText(STAMP stamp)
 	pt.y = r.corner.y + SP_Y_PADDING;
 
 	s.frame = SetAbsFrameIndex(MeleeFrame,
-							   SPECIAL_PC + (optControllerType * 4));
+							   SPECIAL_PC + (static_cast<int>(optControllerType) * 4));
 
 	GetFrameRect(s.frame, &r);
 	pt.x -= r.extent.width;
@@ -573,7 +573,7 @@ DrawTeamPickerText(STAMP stamp)
 	pt.y = r.corner.y + r.extent.height - TP_PADDING;
 
 	s.frame = SetAbsFrameIndex(MeleeFrame,
-							   CONFIRM_PC + (optControllerType * 4));
+							   CONFIRM_PC + (static_cast<int>(optControllerType) * 4));
 
 	GetFrameRect(s.frame, &r);
 	pt.y -= r.extent.height;
@@ -602,7 +602,7 @@ DrawTeamPickerText(STAMP stamp)
 		}
 
 		s.frame = SetAbsFrameIndex(MeleeFrame,
-								   CANCEL_PC + i + (optControllerType * 4));
+								   CANCEL_PC + i + (static_cast<int>(optControllerType) * 4));
 		s.origin.x = text_r.extent.width + t.baseline.x + TP_PADDING
 				   + RES_SCALE(1);
 		DrawStamp(&s);
